# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An `API` for cross-platform custom orchestration of execution steps\nwithout any third-party dependencies. Based on `DAG`, it implements the scheduling\nfunction of sequential execution of dependent steps and concurrent execution of\nnon-dependent steps. \u003cbr /\u003e\u003cbr /\u003e It provides `API` remote operation mode, batch\nexecution of `Shell` , `Powershell` , `Python` and other commands, and easily\ncompletes common management tasks such as running automated operation and maintenance\nscripts, polling processes, installing or uninstalling software, updating applications,\nand installing patches.",
        "title": "DAG Flow",
        "contact": {
            "name": "dagflow",
            "url": "https://github.com/busyster996/dagflow/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/busyster996/dagflow/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/event": {
            "get": {
                "description": "订阅系统事件, 仅支持SSE订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "事件"
                ],
                "summary": "事件",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline": {
            "get": {
                "description": "获取所有流水线列表, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流水线"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "分页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SPipelineListRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "创建流水线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流水线"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "description": "流水线内容",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SPipelineCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline/{pipeline}": {
            "get": {
                "description": "获取指定流水线详情, 支持SSE订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流水线"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SPipelineRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "更新指定流水线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流水线"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SPipelineUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定流水线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流水线"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline/{pipeline}/build": {
            "get": {
                "description": "获取指定流水线构建任务列表, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "构建"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "分页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SPipelineBuildListRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "创建构建任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "构建"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "构建参数",
                        "name": "build",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SPipelineBuildReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_STaskCreateRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/pipeline/{pipeline}/build/{build}": {
            "get": {
                "description": "获取流水线指定构建任务详情, 支持SSE订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "构建"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "构建名称",
                        "name": "build",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SPipelineBuildRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "重新执行指定构建任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "构建"
                ],
                "summary": "重新运行",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "构建名称",
                        "name": "build",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定构建任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "构建"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "流水线名称",
                        "name": "pipeline",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "构建名称",
                        "name": "build",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/pool": {
            "get": {
                "description": "获取工作池信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作池"
                ],
                "summary": "详情, 支持SSE订阅",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SPoolRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "设置工作池大小",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作池"
                ],
                "summary": "设置",
                "parameters": [
                    {
                        "description": "pool setting",
                        "name": "setting",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SPoolReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SPoolReq"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/pty": {
            "get": {
                "description": "Websocket的pty终端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "终端"
                ],
                "summary": "PTY",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/task": {
            "get": {
                "description": "获取任务列表, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "分页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_STaskListDetailRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "创建任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "description": "任务内容",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.STaskReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_STaskCreateRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{task}": {
            "get": {
                "description": "获取任务详情, 支持SSE订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_STaskRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "put": {
                "description": "管理任务, 支持暂停、恢复、终止、超时暂停自动恢复",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "paused",
                            "kill",
                            "pause",
                            "resume"
                        ],
                        "type": "string",
                        "default": "paused",
                        "description": "操作项",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "1m",
                        "description": "暂停多久, 如果没设置则需要手工恢复",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{task}/step": {
            "get": {
                "description": "获取指定任务的步骤列表, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "步骤"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SStepsRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/task/{task}/step/{step}/log": {
            "get": {
                "description": "指定任务指定步骤的执行输出, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "步骤"
                ],
                "summary": "日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "步骤名称",
                        "name": "step",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SStepLogsRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
...
        "/heartbeat": {
            "get": {
                "description": "用于判断服务是否正常",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "默认"
                ],
                "summary": "心跳",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.SBase-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "types.SBase-types_SPipelineListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.SPipelineListRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_SPipelineRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.SPipelineRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
    }
}